Add files via upload
Objective of our project was to create system which
maintains the stock of medicines and map it with the
suppliers. It is operated by the owner of the medical
and staff as well.
Following activities are covered through our
application:
Owner can add, search, remove and update
information about medicine.
Owner can easily track daily purchases of medicine.